Fix id conflict in manga list and manga disparition.
If two mangas are detect as the same manga by anylist, the anylist id
would be used for both mangas, resulting in a listing error as having
two items with the same id is not supported in a x-for.

This also addresses the case where anylist returns an error (e.g.: when
getting rate-limited), in which case the previous implementation would
simply make the manga disappear. It will now be shown without the
enhanced information.